+++ src/fmt.rs
@@ -4,7 +4,6 @@
 #[cfg(all(feature = "defmt", feature = "log"))]
 compile_error!("You may not enable both `defmt` and `log` features.");
 
-
 macro_rules! trace {
     ($s:literal $(, $x:expr)* $(,)?) => {
         {
@@ -30,4 +29,3 @@ macro_rules! debug {
         }
     };
 }
-