+++ src/protocol/rc5/tests.rs
@@ -1,7 +1,9 @@
-use crate::protocol::rc5::Rc5Command;
-use crate::protocol::Rc5;
-use crate::receiver::{Builder, Receiver};
-use crate::sender::PulsedataBuffer;
+use crate::{
+    protocol::{rc5::Rc5Command, Rc5},
+    receiver::Builder,
+    remotecontrol::Action,
+    sender::PulsedataBuffer,
+};
 
 #[test]
 fn rc5_command() {
@@ -16,7 +18,7 @@ fn test_bufrecv() {
         72, 72, 73, 70, 72, 36, 37, 34, 36, 36, 36, 71, 73, 35, 37, 70, 37,
     ];
 
-    let mut r = Receiver::builder()
+    let mut r = Builder::new()
         .rc5()
         .resolution(40_000)
         .buffer(&dists)
@@ -33,6 +35,9 @@ fn test_bufrecv() {
 #[test]
 #[rustfmt::skip]
 fn command_mixed() {
+
+    use crate::remotecontrol::rc5;
+
     let dists = [
         57910, 36, 36, 36, 35, 37, 35, 72, 71, 72, 36, 36, 36, 36, 35, 36, 36, 36, 35, 36, 36, 36, 71, 36,
         26605, 36, 36, 71, 72, 72, 71, 72, 36, 36, 35, 36, 36, 36, 35, 37, 35, 36, 36, 36, 71, 36,
@@ -41,14 +46,30 @@ fn command_mixed() {
         50973, 38, 34, 73, 70, 73, 70, 74, 34, 37, 35, 37, 34, 38, 34, 37, 35, 37, 34, 38, 70, 37,
     ];
 
-    let mut r = Builder::<Rc5>::new().resolution(40_000).buffer(&dists).build();
+    let mut r = Builder::new()
+        .rc5()
+        .resolution(40_000).buffer(&dists).build();
     let v: std::vec::Vec<_> = r.iter().collect();
     assert_eq!(v.len(), 3);
 
     for c in &v {
+        println!("{:?}", c);
         assert_eq!(c.addr, 20);
         assert_eq!(c.cmd, 1);
     }
+
+    let mut r = crate::Receiver::builder()
+        .rc5()
+        .remotecontrol(rc5::CdPlayer)
+        .resolution(40_000).buffer(&dists).build();
+
+    let v: std::vec::Vec<_> = r.iter().collect();
+    assert_eq!(v.len(), 3);
+
+    for c in &v {
+        assert_eq!(c.action(), Some(Action::One));
+    }
+
 }
 
 #[test]
@@ -63,12 +84,14 @@ fn all_commands() {
 
             let cmd: Rc5Command = Rc5Command::new(address, cmdnum, false);
             ptb.load::<Rc5, SAMPLERATE>(&cmd);
-            let mut r = Builder::<Rc5>::new()
+            let mut r = Builder::new()
+                .rc5()
                 .resolution(SAMPLERATE)
                 .buffer(&ptb.buf)
                 .build();
 
             let cmdres = r.iter().next().unwrap();
+
             assert_eq!(cmd.addr, cmdres.addr);
             assert_eq!(cmd.cmd, cmdres.cmd);
         }
@@ -77,16 +100,16 @@ fn all_commands() {
 
 #[test]
 fn timer_resolution() {
-    one_freq::<20_000>();
-    one_freq::<40_000>();
-    one_freq::<48_000_000>();
+    test_freq::<20_000>();
+    test_freq::<40_000>();
+    test_freq::<48_000_000>();
 }
 
-fn one_freq<const F: usize>() {
+fn test_freq<const F: usize>() {
     let mut ptb = PulsedataBuffer::<96>::new();
     let cmd: Rc5Command = Rc5Command::new(10, 2, false);
     ptb.load::<Rc5, F>(&cmd);
-    let mut r = Builder::<Rc5>::new().resolution(F).buffer(&ptb.buf).build();
+    let mut r = Builder::new().rc5().resolution(F).buffer(&ptb.buf).build();
 
     let cmdres = r.iter().next().unwrap();
     assert_eq!(cmd.addr, cmdres.addr);