+++ src/remotecontrol/sbp/blurayplayer.rs
@@ -1,18 +1,21 @@
-use crate::protocol::SbpCommand;
-use crate::remotecontrol::{Button, DeviceType, RemoteControl};
-use crate::ProtocolId;
+use crate::{
+    protocol::SbpCommand,
+    remotecontrol::{Action, DeviceType, RemoteControlModel},
+    ProtocolId,
+};
 
-use Button::*;
+use Action::*;
 
+#[derive(Debug, Default)]
 pub struct SamsungBluRayPlayer;
 
-impl RemoteControl for SamsungBluRayPlayer {
+impl RemoteControlModel for SamsungBluRayPlayer {
     const MODEL: &'static str = "Samsung BluRay Player";
     const DEVTYPE: DeviceType = DeviceType::BluRayPlayer;
     const PROTOCOL: ProtocolId = ProtocolId::Sbp;
     const ADDRESS: u32 = 32;
     type Cmd = SbpCommand;
-    const BUTTONS: &'static [(u32, Button)] = &[
+    const BUTTONS: &'static [(u32, Action)] = &[
         (2, One),
         (3, Two),
         (4, Three),