+++ tests/builder.rs
@@ -1,13 +1,16 @@
 use embedded_hal::digital::v2::InputPin;
-use infrared::protocol::NecSamsung;
-use infrared::protocol::Rc5;
-use infrared::receiver::{Builder, ConstReceiver, Event, Poll, Receiver};
+use infrared::{
+    protocol::{NecSamsung, Rc5},
+    receiver::{ConstReceiver, DefaultInput, Event, Poll, Receiver},
+    remotecontrol::{rc5::CdPlayer, Button},
+};
 
 #[test]
 fn const_embedded_hal_constreceiver() {
     let pin = DummyEmbeddedHalPin;
 
-    let mut recv = Builder::<Rc5>::new()
+    let mut recv = Receiver::builder()
+        .rc5()
         .polled()
         .pin(pin)
         .build_const::<1_000_000>();
@@ -15,8 +18,11 @@ fn const_embedded_hal_constreceiver() {
     recv.poll().unwrap();
 
     let pin = DummyEmbeddedHalPin;
-    let mut recv: ConstReceiver<Rc5, Event, _, 1_000_000> =
-        Builder::<Rc5>::new().event_driven().pin(pin).build_const();
+    let mut recv: ConstReceiver<Rc5, Event, _, 1_000_000> = Receiver::builder()
+        .rc5()
+        .event_driven()
+        .pin(pin)
+        .build_const();
 
     let _ = recv.event(100);
 }
@@ -25,7 +31,8 @@ fn const_embedded_hal_constreceiver() {
 fn const_embedded_hal_receiver() {
     let pin = DummyEmbeddedHalPin;
 
-    let mut recv: Receiver<Rc5, Poll, _> = Builder::<Rc5>::new()
+    let mut recv: Receiver<Rc5, Poll, _> = Receiver::builder()
+        .rc5()
         .polled()
         .resolution(20_000)
         .pin(pin)
@@ -34,11 +41,14 @@ fn const_embedded_hal_receiver() {
     let _ = recv.poll();
 
     let pin = DummyEmbeddedHalPin;
-    let mut recv: Receiver<Rc5, Event, _> = Builder::<Rc5>::new()
+    let mut recv = Receiver::builder()
+        .rc5()
+        .remotecontrol(CdPlayer)
         .event_driven()
         .resolution(20_000)
         .pin(pin)
         .build();
+
     let _ = recv.event(204);
 
     let _recv = Receiver::builder()
@@ -60,13 +70,12 @@ fn receiver_iterator() {
 
     // Iterate through the commands
     let _i = recv.iter();
-
-    //let mut r: BufferReceiver<Rc5> = BufferReceiver::with_buf(20_000, &[]);
 }
 
 #[test]
 fn receiver_generic() {
-    let mut recv: Receiver<Rc5, Event, _> = Builder::<Rc5>::new()
+    let mut recv: Receiver<Rc5, Event, _> = Receiver::builder()
+        .rc5()
         .event_driven()
         .resolution(20_000)
         .build();
@@ -80,6 +89,39 @@ fn receiver_generic() {
         .build();
     receiver.poll();
      */
+}
+
+#[test]
+fn receiver_remote() {
+    use infrared::remotecontrol::rc5;
+
+    let mut r: Receiver<Rc5, Event, DefaultInput, Button<CdPlayer>> = Receiver::builder()
+        .event_driven()
+        .rc5()
+        .remotecontrol(rc5::CdPlayer)
+        .build();
+
+    match r.event(40, false) {
+        Ok(Some(event)) => {
+            println!(
+                "Action: {:?}, repeat: {}",
+                event.action(),
+                event.is_repeat()
+            )
+        }
+        Ok(None) => (),
+        Err(_err) => (),
+    }
+
+    let _r: Receiver<Rc5, Event, DefaultInput, Button<CdPlayer>> = Receiver::new(20_000);
+
+    let _r: Receiver<Rc5, Poll, DefaultInput, Button<CdPlayer>> = Receiver::new(20_000);
+
+    let _r: Receiver<Rc5, Poll, DefaultInput, Button<CdPlayer>> = Receiver::builder()
+        .rc5()
+        .polled()
+        .remotecontrol(CdPlayer)
+        .build();
 }
 
 struct DummyEmbeddedHalPin;